requirements.txt
# requirements.txt
psycopg2-binary==2.9.5
requests==2.28.2
python-dotenv==0.19.2